Modularise cloud functions
